Run multiple dependency tests in a different node
TODO: It is not computing transitive closure and/or reduction, so there
is no dependency added between the "multiple dependency test" and model
following test's upstream dependencies. It should be computed and fixed
in the future.

E.g., for a following model, where M1, M2, and M3 are models, and T1 is
a test with dependency on both M1 and M2:

  M1 -----> M2 -> M3
   \        /
    -> T1 <-

M3 should be run AFTER T1. In the current version, there is no
additional dependency between T1 and M3 created, so they both are run
independently.